[ffhp] Custom theme overrides system; refactor FFHP code
