+++ src/internal/data/base.lua
@@ -4,3 +4,4 @@ require("internal.data.events")
 require("internal.data.config_option_type")
 require("internal.data.config_menu")
 require("internal.data.config_option")
+require("internal.data.theme")
