+++ src/mod/base/data/init.lua
@@ -2,5 +2,4 @@ require("mod.base.data.language")
 require("mod.base.data.doc")
 require("mod.base.data.keybind")
 require("mod.base.data.asset")
-require("mod.base.data.theme")
 require("mod.base.data.auto_turn_anim")
