--- src/mod/base/data/theme.lua
@@ -1,6 +1,0 @@
-data:add {
-   _type = "base.theme",
-   _id = "default",
-
-   overrides = {}
-}
