+++ src/mod/ceri_items/data/chip.lua
@@ -709,1070 +709,3 @@ local spellbook = {
    }
 }
 data:add_multi("base.chip", spellbook)
-
-data:add_type {
-   name = "ffhp_mapping",
-   fields = {
-      {
-         name = "item_id",
-         template = true,
-         default = "",
-         type = "id:base.item"
-      },
-      {
-         name = "chip_on_identify",
-         template = true,
-         default = "",
-         type = "id:base.chip?"
-      },
-      {
-         name = "color_on_identify",
-         template = true,
-         default = {255, 255, 255},
-         type = "color?"
-      }
-   }
-}
-
-local mappings = {
-   {
-      _id = "ex_potion_540",
-      item_id = "elona.potion_of_blindness",
-      chip_on_identify = "ceri_items.ex_potion_540",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_548",
-      item_id = "elona.potion_of_troll_blood",
-      chip_on_identify = "ceri_items.ex_potion_548",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_615",
-      item_id = "elona.spellbook_of_contingency",
-      chip_on_identify = "ceri_items.ex_spellbook_615",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_674",
-      item_id = "elona.flying_scroll",
-      chip_on_identify = "ceri_items.ex_scroll_674",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_552",
-      item_id = "elona.potion_of_resistance",
-      chip_on_identify = "ceri_items.ex_potion_552",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_639",
-      item_id = "elona.spellbook_of_elemental_scar",
-      chip_on_identify = "ceri_items.ex_spellbook_639",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_586",
-      item_id = "elona.rod_of_slow",
-      chip_on_identify = "ceri_items.ex_rod_586",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_666",
-      item_id = "elona.scroll_of_holy_light",
-      chip_on_identify = "ceri_items.ex_scroll_666",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_533",
-      item_id = "elona.potion_of_cure_major_wound",
-      chip_on_identify = "ceri_items.ex_potion_533",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_529",
-      item_id = "elona.potion_of_evolution",
-      chip_on_identify = "ceri_items.ex_potion_529",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_612",
-      item_id = "elona.spellbook_of_speed",
-      chip_on_identify = "ceri_items.ex_spellbook_612",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_ffhp_686",
-      item_id = "elona.treasure_map",
-      chip_on_identify = "ceri_items.ex_ffhp_686",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_614",
-      item_id = "elona.spellbook_of_holy_veil",
-      chip_on_identify = "ceri_items.ex_spellbook_614",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_547",
-      item_id = "elona.potion_of_speed",
-      chip_on_identify = "ceri_items.ex_potion_547",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_657",
-      item_id = "elona.scroll_of_identify",
-      chip_on_identify = "ceri_items.ex_scroll_657",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_551",
-      item_id = "elona.potion_of_restore_spirit",
-      chip_on_identify = "ceri_items.ex_potion_551",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_606",
-      item_id = "elona.spellbook_of_holy_light",
-      chip_on_identify = "ceri_items.ex_spellbook_606",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_649",
-      item_id = "elona.scroll_of_curse",
-      chip_on_identify = "ceri_items.ex_scroll_649",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_646",
-      item_id = "elona.spellbook_of_harvest",
-      chip_on_identify = "ceri_items.ex_spellbook_646",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_520",
-      item_id = "elona.bottle_of_water",
-      chip_on_identify = "ceri_items.ex_potion_520",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_623",
-      item_id = "elona.spellbook_of_darkness_arrow",
-      chip_on_identify = "ceri_items.ex_spellbook_623",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_528",
-      item_id = "elona.potion_of_potential",
-      chip_on_identify = "ceri_items.ex_potion_528",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_ffhp_685",
-      item_id = "elona.bejeweled_chest",
-      chip_on_identify = "ceri_items.ex_ffhp_685",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_532",
-      item_id = "elona.potion_of_cure_minor_wound",
-      chip_on_identify = "ceri_items.ex_potion_532",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_609",
-      item_id = "elona.spellbook_of_holy_shield",
-      chip_on_identify = "ceri_items.ex_spellbook_609",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_668",
-      item_id = "elona.scroll_of_faith",
-      chip_on_identify = "ceri_items.ex_scroll_668",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_625",
-      item_id = "elona.spellbook_of_fire_bolt",
-      chip_on_identify = "ceri_items.ex_spellbook_625",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_676",
-      item_id = "elona.scroll_of_change_material",
-      chip_on_identify = "ceri_items.ex_scroll_676",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_652",
-      item_id = "elona.scroll_of_return",
-      chip_on_identify = "ceri_items.ex_scroll_652",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_ffhp_688",
-      item_id = "elona.happy_bed",
-      chip_on_identify = "ceri_items.ex_ffhp_688",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_595",
-      item_id = "elona.spellbook_of_nightmare",
-      chip_on_identify = "ceri_items.ex_spellbook_595",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_577",
-      item_id = "elona.rod_of_web",
-      chip_on_identify = "ceri_items.ex_rod_577",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_ffhp_684",
-      item_id = "elona.chest",
-      chip_on_identify = "ceri_items.ex_ffhp_684",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_681",
-      item_id = "elona.scroll_of_greater_enchant_armor",
-      chip_on_identify = "ceri_items.ex_scroll_681",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_689",
-      item_id = "elona.scroll_of_uncurse",
-      chip_on_identify = "ceri_items.ex_scroll_689",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_565",
-      item_id = "elona.rod_of_change_creature",
-      chip_on_identify = "ceri_items.ex_rod_565",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_592",
-      item_id = "elona.spellbook_of_magic_mapping",
-      chip_on_identify = "ceri_items.ex_spellbook_592",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_624",
-      item_id = "elona.spellbook_of_ice_bolt",
-      chip_on_identify = "ceri_items.ex_spellbook_624",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_675",
-      item_id = "elona.scroll_of_inferior_material",
-      chip_on_identify = "ceri_items.ex_scroll_675",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_600",
-      item_id = "elona.spellbook_of_cure_minor_wound",
-      chip_on_identify = "ceri_items.ex_spellbook_600",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_659",
-      item_id = "elona.scroll_of_mana",
-      chip_on_identify = "ceri_items.ex_scroll_659",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_640",
-      item_id = "elona.spellbook_of_acid_ground",
-      chip_on_identify = "ceri_items.ex_spellbook_640",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_641",
-      item_id = "elona.spellbook_of_fire_wall",
-      chip_on_identify = "ceri_items.ex_spellbook_641",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_570",
-      item_id = "elona.rod_of_cure_minor_wound",
-      chip_on_identify = "ceri_items.ex_rod_570",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_582",
-      item_id = "elona.rod_of_lightning_bolt",
-      chip_on_identify = "ceri_items.ex_rod_582",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_680",
-      item_id = "elona.scroll_of_enchant_armor",
-      chip_on_identify = "ceri_items.ex_scroll_680",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_619",
-      item_id = "elona.spellbook_of_magic_arrow",
-      chip_on_identify = "ceri_items.ex_spellbook_619",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_670",
-      item_id = "elona.scroll_of_recharge",
-      chip_on_identify = "ceri_items.ex_scroll_670",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_544",
-      item_id = "elona.potion_of_weakness",
-      chip_on_identify = "ceri_items.ex_potion_544",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_576",
-      item_id = "elona.rod_of_uncurse",
-      chip_on_identify = "ceri_items.ex_rod_576",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_556",
-      item_id = "elona.fireproof_liquid",
-      chip_on_identify = "ceri_items.ex_potion_556",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_654",
-      item_id = "elona.scroll_of_healing_rain",
-      chip_on_identify = "ceri_items.ex_scroll_654",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_662",
-      item_id = "elona.scroll_of_gain_attribute",
-      chip_on_identify = "ceri_items.ex_scroll_662",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_635",
-      item_id = "elona.spellbook_of_wall_creation",
-      chip_on_identify = "ceri_items.ex_spellbook_635",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_573",
-      item_id = "elona.rod_of_heal",
-      chip_on_identify = "ceri_items.ex_rod_573",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_537",
-      item_id = "elona.potion_of_healer_eris",
-      chip_on_identify = "ceri_items.ex_potion_537",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_569",
-      item_id = "elona.rod_of_fire_wall",
-      chip_on_identify = "ceri_items.ex_rod_569",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_597",
-      item_id = "elona.spellbook_of_incognito",
-      chip_on_identify = "ceri_items.ex_spellbook_597",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_525",
-      item_id = "elona.potion_of_cure_corruption",
-      chip_on_identify = "ceri_items.ex_potion_525",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_571",
-      item_id = "elona.rod_of_cure",
-      chip_on_identify = "ceri_items.ex_rod_571",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_567",
-      item_id = "elona.rod_of_wall_creation",
-      chip_on_identify = "ceri_items.ex_rod_567",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_632",
-      item_id = "elona.spellbook_of_chaos_ball",
-      chip_on_identify = "ceri_items.ex_spellbook_632",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_578",
-      item_id = "elona.rod_of_summon_monsters",
-      chip_on_identify = "ceri_items.ex_rod_578",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_ffhp_692",
-      item_id = "elona.raw_noodle",
-      chip_on_identify = "ceri_items.ex_ffhp_692",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_653",
-      item_id = "elona.scroll_of_escape",
-      chip_on_identify = "ceri_items.ex_scroll_653",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_661",
-      item_id = "elona.scroll_of_growth",
-      chip_on_identify = "ceri_items.ex_scroll_661",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_669",
-      item_id = "elona.scroll_of_ally",
-      chip_on_identify = "ceri_items.ex_scroll_669",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_555",
-      item_id = "elona.acidproof_liquid",
-      chip_on_identify = "ceri_items.ex_potion_555",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_610",
-      item_id = "elona.spellbook_of_regeneration",
-      chip_on_identify = "ceri_items.ex_spellbook_610",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_626",
-      item_id = "elona.spellbook_of_lightning_bolt",
-      chip_on_identify = "ceri_items.ex_spellbook_626",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_603",
-      item_id = "elona.spellbook_of_cure_jure",
-      chip_on_identify = "ceri_items.ex_spellbook_603",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_611",
-      item_id = "elona.spellbook_of_resistance",
-      chip_on_identify = "ceri_items.ex_spellbook_611",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_572",
-      item_id = "elona.rod_of_healing_hands",
-      chip_on_identify = "ceri_items.ex_rod_572",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_596",
-      item_id = "elona.spellbook_of_mutation",
-      chip_on_identify = "ceri_items.ex_spellbook_596",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_524",
-      item_id = "elona.love_potion",
-      chip_on_identify = "ceri_items.ex_potion_524",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_643",
-      item_id = "elona.spellbook_of_magic_laser",
-      chip_on_identify = "ceri_items.ex_spellbook_643",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_536",
-      item_id = "elona.potion_of_healer_odina",
-      chip_on_identify = "ceri_items.ex_potion_536",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_638",
-      item_id = "elona.spellbook_of_weakness",
-      chip_on_identify = "ceri_items.ex_spellbook_638",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_637",
-      item_id = "elona.spellbook_of_slow",
-      chip_on_identify = "ceri_items.ex_spellbook_637",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_636",
-      item_id = "elona.spellbook_of_wishing",
-      chip_on_identify = "ceri_items.ex_spellbook_636",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_634",
-      item_id = "elona.spellbook_of_web",
-      chip_on_identify = "ceri_items.ex_spellbook_634",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_633",
-      item_id = "elona.spellbook_of_sound_ball",
-      chip_on_identify = "ceri_items.ex_spellbook_633",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_631",
-      item_id = "elona.spellbook_of_fire_ball",
-      chip_on_identify = "ceri_items.ex_spellbook_631",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_630",
-      item_id = "elona.spellbook_of_ice_ball",
-      chip_on_identify = "ceri_items.ex_spellbook_630",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_629",
-      item_id = "elona.spellbook_of_summon_monsters",
-      chip_on_identify = "ceri_items.ex_spellbook_629",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_627",
-      item_id = "elona.spellbook_of_darkness_beam",
-      chip_on_identify = "ceri_items.ex_spellbook_627",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_622",
-      item_id = "elona.spellbook_of_chaos_eye",
-      chip_on_identify = "ceri_items.ex_spellbook_622",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_621",
-      item_id = "elona.spellbook_of_nerve_eye",
-      chip_on_identify = "ceri_items.ex_spellbook_621",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_618",
-      item_id = "elona.spellbook_of_oracle",
-      chip_on_identify = "ceri_items.ex_spellbook_618",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_617",
-      item_id = "elona.spellbook_of_minor_teleportation",
-      chip_on_identify = "ceri_items.ex_spellbook_617",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_616",
-      item_id = "elona.spellbook_of_teleportation",
-      chip_on_identify = "ceri_items.ex_spellbook_616",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_613",
-      item_id = "elona.spellbook_of_hero",
-      chip_on_identify = "ceri_items.ex_spellbook_613",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_608",
-      item_id = "elona.spellbook_of_uncurse",
-      chip_on_identify = "ceri_items.ex_spellbook_608",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_607",
-      item_id = "elona.spellbook_of_holy_rain",
-      chip_on_identify = "ceri_items.ex_spellbook_607",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_601",
-      item_id = "elona.spellbook_of_cure_critical_wound",
-      chip_on_identify = "ceri_items.ex_spellbook_601",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_598",
-      item_id = "elona.spellbook_of_4_dimensional_pocket",
-      chip_on_identify = "ceri_items.ex_spellbook_598",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_594",
-      item_id = "elona.spellbook_of_silence",
-      chip_on_identify = "ceri_items.ex_spellbook_594",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_550",
-      item_id = "elona.potion_of_restore_body",
-      chip_on_identify = "ceri_items.ex_potion_550",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_593",
-      item_id = "elona.spellbook_of_detect_objects",
-      chip_on_identify = "ceri_items.ex_spellbook_593",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_589",
-      item_id = "elona.ancient_book",
-      chip_on_identify = "ceri_items.ex_spellbook_589",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_682",
-      item_id = "elona.book_of_resurrection",
-      chip_on_identify = "ceri_items.ex_scroll_682",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_584",
-      item_id = "elona.rod_of_acid_ground",
-      chip_on_identify = "ceri_items.ex_rod_584",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_679",
-      item_id = "elona.scroll_of_greater_enchant_weapon",
-      chip_on_identify = "ceri_items.ex_scroll_679",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_678",
-      item_id = "elona.scroll_of_enchant_weapon",
-      chip_on_identify = "ceri_items.ex_scroll_678",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_677",
-      item_id = "elona.scroll_of_superior_material",
-      chip_on_identify = "ceri_items.ex_scroll_677",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_673",
-      item_id = "elona.scroll_of_name",
-      chip_on_identify = "ceri_items.ex_scroll_673",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_672",
-      item_id = "elona.scroll_of_contingency",
-      chip_on_identify = "ceri_items.ex_scroll_672",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_667",
-      item_id = "elona.scroll_of_holy_rain",
-      chip_on_identify = "ceri_items.ex_scroll_667",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_665",
-      item_id = "elona.scroll_of_holy_veil",
-      chip_on_identify = "ceri_items.ex_scroll_665",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_605",
-      item_id = "elona.spellbook_of_healing_hands",
-      chip_on_identify = "ceri_items.ex_spellbook_605",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_663",
-      item_id = "elona.scroll_of_wonder",
-      chip_on_identify = "ceri_items.ex_scroll_663",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_660",
-      item_id = "elona.scroll_of_knowledge",
-      chip_on_identify = "ceri_items.ex_scroll_660",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_658",
-      item_id = "elona.scroll_of_greater_identify",
-      chip_on_identify = "ceri_items.ex_scroll_658",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_651",
-      item_id = "elona.scroll_of_teleportation",
-      chip_on_identify = "ceri_items.ex_scroll_651",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_650",
-      item_id = "elona.scroll_of_minor_teleportation",
-      chip_on_identify = "ceri_items.ex_scroll_650",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_539",
-      item_id = "elona.potion_of_healing",
-      chip_on_identify = "ceri_items.ex_potion_539",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_585",
-      item_id = "elona.rod_of_speed",
-      chip_on_identify = "ceri_items.ex_rod_585",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_583",
-      item_id = "elona.rod_of_magic_missile",
-      chip_on_identify = "ceri_items.ex_rod_583",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_581",
-      item_id = "elona.rod_of_fire_bolt",
-      chip_on_identify = "ceri_items.ex_rod_581",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_580",
-      item_id = "elona.rod_of_ice_bolt",
-      chip_on_identify = "ceri_items.ex_rod_580",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_579",
-      item_id = "elona.rod_of_silence",
-      chip_on_identify = "ceri_items.ex_rod_579",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_575",
-      item_id = "elona.rod_of_holy_light",
-      chip_on_identify = "ceri_items.ex_rod_575",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_574",
-      item_id = "elona.rod_of_mana",
-      chip_on_identify = "ceri_items.ex_rod_574",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_664",
-      item_id = "elona.scroll_of_gain_material",
-      chip_on_identify = "ceri_items.ex_scroll_664",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_564",
-      item_id = "elona.rod_of_domination",
-      chip_on_identify = "ceri_items.ex_rod_564",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_566",
-      item_id = "elona.rod_of_alchemy",
-      chip_on_identify = "ceri_items.ex_rod_566",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_562",
-      item_id = "elona.rod_of_teleportation",
-      chip_on_identify = "ceri_items.ex_rod_562",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_560",
-      item_id = "elona.rod_of_identify",
-      chip_on_identify = "ceri_items.ex_rod_560",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_554",
-      item_id = "elona.potion_of_hero",
-      chip_on_identify = "ceri_items.ex_potion_554",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_553",
-      item_id = "elona.potion_of_defender",
-      chip_on_identify = "ceri_items.ex_potion_553",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_549",
-      item_id = "elona.bottle_of_hermes_blood",
-      chip_on_identify = "ceri_items.ex_potion_549",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_543",
-      item_id = "elona.potion_of_silence",
-      chip_on_identify = "ceri_items.ex_potion_543",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_542",
-      item_id = "elona.potion_of_paralysis",
-      chip_on_identify = "ceri_items.ex_potion_542",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_541",
-      item_id = "elona.potion_of_confusion",
-      chip_on_identify = "ceri_items.ex_potion_541",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_648",
-      item_id = "elona.scroll_of_vanish_curse",
-      chip_on_identify = "ceri_items.ex_scroll_648",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_546",
-      item_id = "elona.potion_of_slow",
-      chip_on_identify = "ceri_items.ex_potion_546",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_645",
-      item_id = "elona.spellbook_of_domination",
-      chip_on_identify = "ceri_items.ex_spellbook_645",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_535",
-      item_id = "elona.potion_of_healer",
-      chip_on_identify = "ceri_items.ex_potion_535",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_534",
-      item_id = "elona.potion_of_cure_critical_wound",
-      chip_on_identify = "ceri_items.ex_potion_534",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_531",
-      item_id = "elona.poison",
-      chip_on_identify = "ceri_items.ex_potion_531",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_527",
-      item_id = "elona.potion_of_cure_mutation",
-      chip_on_identify = "ceri_items.ex_potion_527",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_523",
-      item_id = "elona.bottle_of_dye",
-      chip_on_identify = "ceri_items.ex_potion_523",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_522",
-      item_id = "elona.empty_bottle",
-      chip_on_identify = "ceri_items.ex_potion_522",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_521",
-      item_id = "elona.bottle_of_dirty_water",
-      chip_on_identify = "ceri_items.ex_potion_521",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_ffhp_693",
-      item_id = "elona.heir_trunk",
-      chip_on_identify = "ceri_items.ex_ffhp_693",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_ffhp_691",
-      item_id = "elona.ration",
-      chip_on_identify = "ceri_items.ex_ffhp_691",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_ffhp_690",
-      item_id = "elona.music_ticket",
-      chip_on_identify = "ceri_items.ex_ffhp_690",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_ffhp_687",
-      item_id = "elona.sisters_love_fueled_lunch",
-      chip_on_identify = "ceri_items.ex_ffhp_687",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_ffhp_587",
-      item_id = "elona.cat_sisters_diary",
-      chip_on_identify = "ceri_items.ex_ffhp_587",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_ffhp_558",
-      item_id = "elona.little_sisters_diary",
-      chip_on_identify = "ceri_items.ex_ffhp_558",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_656",
-      item_id = "elona.scroll_of_detect_objects",
-      chip_on_identify = "ceri_items.ex_scroll_656",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_599",
-      item_id = "elona.spellbook_of_knowledge",
-      chip_on_identify = "ceri_items.ex_spellbook_599",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_ffhp_559",
-      item_id = "elona.girls_diary",
-      chip_on_identify = "ceri_items.ex_ffhp_559",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_563",
-      item_id = "elona.rod_of_wishing",
-      chip_on_identify = "ceri_items.ex_rod_563",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_602",
-      item_id = "elona.spellbook_of_cure_eris",
-      chip_on_identify = "ceri_items.ex_spellbook_602",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_561",
-      item_id = "elona.rod_of_magic_mapping",
-      chip_on_identify = "ceri_items.ex_rod_561",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_642",
-      item_id = "elona.spellbook_of_make_door",
-      chip_on_identify = "ceri_items.ex_spellbook_642",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_604",
-      item_id = "elona.spellbook_of_healing_rain",
-      chip_on_identify = "ceri_items.ex_spellbook_604",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_557",
-      item_id = "elona.potion_of_descent",
-      chip_on_identify = "ceri_items.ex_potion_557",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_628",
-      item_id = "elona.spellbook_of_illusion_beam",
-      chip_on_identify = "ceri_items.ex_spellbook_628",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_rod_568",
-      item_id = "elona.rod_of_make_door",
-      chip_on_identify = "ceri_items.ex_rod_568",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_671",
-      item_id = "elona.scroll_of_incognito",
-      chip_on_identify = "ceri_items.ex_scroll_671",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_647",
-      item_id = "elona.scroll_of_oracle",
-      chip_on_identify = "ceri_items.ex_scroll_647",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_644",
-      item_id = "elona.spellbook_of_magic_ball",
-      chip_on_identify = "ceri_items.ex_spellbook_644",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_545",
-      item_id = "elona.potion_of_weaken_resistance",
-      chip_on_identify = "ceri_items.ex_potion_545",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_scroll_655",
-      item_id = "elona.scroll_of_magical_map",
-      chip_on_identify = "ceri_items.ex_scroll_655",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_620",
-      item_id = "elona.spellbook_of_nether_eye",
-      chip_on_identify = "ceri_items.ex_spellbook_620",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_530",
-      item_id = "elona.sleeping_drug",
-      chip_on_identify = "ceri_items.ex_potion_530",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_538",
-      item_id = "elona.potion_of_healer_jure",
-      chip_on_identify = "ceri_items.ex_potion_538",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_590",
-      item_id = "elona.spellbook_of_identify",
-      chip_on_identify = "ceri_items.ex_spellbook_590",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_potion_526",
-      item_id = "elona.potion_of_mutation",
-      chip_on_identify = "ceri_items.ex_potion_526",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_ffhp_683",
-      item_id = "elona.dead_fish",
-      chip_on_identify = "ceri_items.ex_ffhp_683",
-      color_on_identify = {255, 255, 255}
-   },
-   {
-      _id = "ex_spellbook_591",
-      item_id = "elona.spellbook_of_return",
-      chip_on_identify = "ceri_items.ex_spellbook_591",
-      color_on_identify = {255, 255, 255}
-   }
-}
-
-data:add_multi("ceri_items.ffhp_mapping", mappings)
