+++ src/mod/ceri_items/mod.lua
@@ -2,6 +2,7 @@ return {
    id = "ceri_items",
    version = "0.1.0",
    dependencies = {
-      elona = ">= 0"
+      elona = ">= 0",
+      ffhp = ">= 0"
    }
 }
