+++ src/mod/elona_sys/api/Compat.lua
@@ -1,4 +1,5 @@
 local Compat = {}
+local Enum = require("api.Enum")
 
 function Compat.convert_122_id(_type, elona_id)
    local it = data[_type]:find_by("elona_id", elona_id)
@@ -40,6 +41,39 @@ end
 function Compat.convert_122_item_category(category)
    assert(category)
    return data["base.item_type"]:iter():filter(function(i) return i.ordering == category end):extract("_id"):nth(1)
+end
+
+local COLORS = {
+   [0] = "White",
+   [1] = "White",
+   [2] = "Green",
+   [3] = "Red",
+   [4] = "Blue",
+   [5] = "Yellow",
+   [6] = "Brown",
+   [7] = "Black",
+   [8] = "Purple",
+   [9] = "SkyBlue",
+   [10] = "Pink",
+   [11] = "Orange",
+   [12] = "White",
+   [13] = "Fresh",
+   [14] = "DarkGreen",
+   [15] = "Gray",
+   [16] = "LightRed",
+   [17] = "LightBlue",
+   [18] = "LightPurple",
+   [19] = "LightGreen",
+   [20] = "Talk",
+   -- [21] = "RandomFurniture",
+   -- [22] = "RandomSeeded",
+   -- [23] = "RandomAny"
+}
+
+function Compat.convert_122_color_index(color_idx)
+   local color = COLORS[color_idx]
+   assert(color, "Unknown color " .. color_idx)
+   return Enum.Color[color]
 end
 
 return Compat
