+++ src/mod/ffhp/data/init.lua
@@ -1,0 +1,1 @@
+require("mod.ffhp.data.item_ex_mapping")
