+++ src/mod/ffhp/event/init.lua
@@ -1,0 +1,1 @@
+require("mod.ffhp.event.item_ex_mapping")
