+++ src/mod/ffhp/init.lua
@@ -1,0 +1,2 @@
+require("mod.ffhp.data.init")
+require("mod.ffhp.event.init")
