+++ src/mod/ffhp/mod.lua
@@ -1,0 +1,7 @@
+return {
+   id = "ffhp",
+   version = "0.1.0",
+   dependencies = {
+      elona = ">= 0.1.0"
+   }
+}
